+++ cppcache/src/FarSideEntryOp.cpp
@@ -158,7 +158,7 @@ void FarSideEntryOp::skipFilterRoutingInfo(DataInput& input) {
     LOGERROR(
         "FarSideEntryOp::skipFilterRoutingInfo Unexpected type id: %d while "
         "desirializing commit response",
-        structType);
+        static_cast<int32_t>(structType));
     GfErrTypeThrowException(
         "FarSideEntryOp::skipFilterRoutingInfo Unable to handle commit "
         "response",
